Committing back some minor unrelated changes.
 - grok/oldschool/SystemsProg/pidInfo.c
 - grok/oldSchool/README.md
 - web/homepage.html

+++ grok/oldSchool/SystemsProgramming/pidInfo.c
@@ -1,16 +1,4 @@
-/*
- * Print out information on current process.
- *
- *  1. Actually using glib.c wrappers for
- *     underlying Linux versions of these
- *     calls.
- *  2. The unistd.h header seems to be used
- *     oto select a particular POSIX standard.
- *  3. According to the man page, the getpid,
- *     getppid, and getpgrp functions shall
- *     always be successful and no return value
- *     is reserved to indicate an error.
- */
+/* Print out information on current process. */
 
 #include <unistd.h>
 #include <stdio.h>
@@ -18,9 +6,44 @@
 int
 main(void)
 {
-    printf("Process ID: %d\n", getpid());
-    printf("Parent Process ID: %d\n", getppid());
-    printf("Process Group ID: %d\n", getpgrp());
+    printf("Process ID: %ld\n", (long)getpid());
+    printf("Parent Process ID: %ld\n", (long)getppid());
+    printf("Process Group ID: %ld\n", (long)getpgrp());
 
     return 0;
 }
+
+/*
+ * Notes:
+ * 1. The POSIX  Standards do not indicate
+ *    the precise size of pid_t, only that it
+ *    needs to fit inside a long.  For portability,
+ *    we cast the return values of these library
+ *    calls to long.
+ * 2. According to the man page, the getpid,
+ *    getppid, and getpgrp functions shall always
+ *    be successful and no return values are 
+ *    reserved to indicate errors.
+ *
+ * Factoids:
+ * 1. The glibc library is userr-space code that when
+ *    needed calls down into the Linux Kernel.  It does
+ *    this either via system calls or through a shared
+ *    object called the vDSO.  Even actual Linux system
+ *    calls have glibc wrappers for them,
+ * 2. C is the standard defacto ABI for Linux.  Before UNIX,
+ *    the ABI (Application Binary Interface) into an
+ *    operating system was the hardware's assembly language.
+ *    The reason C is such, is that except for isolated uses
+ *    of assembly code, the Unix Kernels are written in C.
+ * 3. Different implementations of C++ are usually binary
+ *    incompatible.  By using 'extern "C" { ... }' directives,
+ *    or foreign function interfaces in other languages,
+ *    C ABI can be the lingua franca that glues code together. 
+ *    Also, allows code to communicate with device drivers
+ *    written in C.
+ *
+ * Suppositions:
+ *  1. The unistd.h header seems to be used
+ *     to select particular POSIX/ISO C standards.
+ */
